// defispace-limit-order/src/contracts/LimitOrder.d.ts
export declare const LimitOrderContract: {
  abi: {
    "ABI version": number;
    header: string[];
    functions: {
      name: string;
      inputs: {name: string; type: string}[];
      outputs: {name: string; type: string}[];
    }[];
    data: {key: number; name: string; type: string}[];
    events: never[];
  };
  tvc: string;
  code: string;
};