// defispace-limit-order/src/contracts/LimitOrderRoot.d.ts
export declare const LimitOrderRootContract: {
  abi: {
    "ABI version": number;
    header: string[];
    functions: {
      name: string;
      inputs: {
        name: string;
        type: string;
      }[];
      outputs: {
        name: string;
        type: string;
      }[];
    }[];
    data: never[];
    events: never[];
  };
  tvc: string;
  code: string;
};